refactor: auth와 president controller 분리
